- fix bug in pafprocess - update experiments readme.

+++ src/pafprocess/pafprocess.h
@@ -53,7 +53,7 @@ struct Connection {
 int process_paf(int p1, int p2, int p3, float *peaks, int h1, int h2, int h3, float *heatmap, int f1, int f2, int f3, float *pafmap);
 int get_num_humans();
 int get_part_cid(int human_id, int part_id);
-int get_score(int human_id);
+float get_score(int human_id);
 int get_part_x(int cid);
 int get_part_y(int cid);
 float get_part_score(int cid);
